+++ framework/mq/pubsub/Queue.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace yii\mq;
+namespace yii\mq\pubsub;
+
+use yii\mq\QueueEvent;
 
 /**
  */
@@ -30,5 +32,21 @@ abstract class Queue extends \yii\base\Component implements QueueInterface
     public function afterPut($message)
 	{
 		$this->trigger(self::EVENT_AFTER_PUT, new QueueEvent(['message'=>$message]));
+	}
+	/**
+	 * @inheritdoc
+	 */
+    public function beforePutSubscription($message, $subscriber_id)
+	{
+		$event = new QueueEvent(['message'=>$message, 'subscriber_id'=>$subscriber_id]);
+		$this->trigger(self::EVENT_BEFORE_PUT_SUBSCRIPTION, $event);
+		return $event->isValid;
+	}
+	/**
+	 * @inheritdoc
+	 */
+    public function afterPutSubscription($message, $subscriber_id)
+	{
+		$this->trigger(self::EVENT_AFTER_PUT_SUBSCRIPTION, new QueueEvent(['message'=>$message, 'subscriber_id'=>$subscriber_id]));
 	}
 }
